More work on setting up base skeleton for work!
Did more work on segregating the testing envrionments for l system trees and shaders.

-Removed the unnecessary view .cpp and .h
-The SupportCanvas3D class can now support an arbitrary amount of sceneview based scenes. It currently supports 2 - one for shaders and one for treees
-Pruned the settings struct
-Made a test scene (xml file) for my shader development

+++ support/scenegraph/Scene.cpp
@@ -73,11 +73,11 @@ void Scene::traverseSceneGraph (Scene *sceneToFill, glm::mat4 accumultedMatrix, 
 
     glm::mat4 accumulatedTransMat = accumultedMatrix * localTransMat;
 
-    for (int i = 0; i < node->primitives.size(); i++){
+    for (unsigned long i = 0; i < node->primitives.size(); i++){
         sceneToFill->addPrimitive(*node->primitives[i], accumulatedTransMat);
     }
 
-    for (int i = 0; i < node->children.size(); i++){
+    for (unsigned long i = 0; i < node->children.size(); i++){
         traverseSceneGraph(sceneToFill, accumulatedTransMat, node->children[i]);
     }
 }
